Make `accessible_role` an `OptionProperty` of every `Widget`
And fix the roles on the accessible widgets in the proof-of-concept to match what I've got in the `Role` enum now
