+++ kivy/core/accessibility/__init__.py
@@ -8,42 +8,6 @@ class AccessibilityBase(object):
     def install(self, window_info):
         pass
 
-class Role(Enum):
-    # Role is the most important property of an accessible widget.
-    # It will tell assistive technologies how to present it to the user, which other properties to expect, and what kind of actions can be performed on it.
-    UNKNOWN = 0
-    """No information about how to present this widget"""
-    CAPTION = 1
-    """This widget's text explains another widget"""
-    GENERIC_CONTAINER = 2
-    """The widget is for containing other widgets
-    
-    In Kivy, generic containers are usually :class:`~kivy.uix.layout.Layout`s.
-    
-    
-    
-    """
-    TOGGLE = 3
-    """The widget is in one of two possible Boolean states
-    
-    Likely either :class:`~kivy.uix.togglebutton.ToggleButton` or
-    :class:`kivy.uix.checkbox.CheckBox`.
-    
-    """
-    BUTTON = 4
-    """A button that does one thing when pressed"""
-    DOCUMENT = 5
-    """A large amount of text the user may want to read
-    
-    DOCUMENT may be a container for many LABEL widgets or similar.
-    
-    """
-    PARAGRAPH = 6
-    """A piece of text within a document"""
-    HEADING = 7
-    """Text starting a section of a document or a menu"""
-    MENU = 8
-    """Modal/drop/accordion/etc. menu with many related widgets inside"""
 
 class Action(Enum):
     # Assistive technologies can request to manipulate widgets on behalf of the user.
